Moved the vessel info rendering to a separate file
