+++ js/app/Visualization/Animation/CartoDBAnimation.js
@@ -3,7 +3,6 @@ define([
   "app/Class",
   "app/LoadingInfo",
   "app/Data/CartoDBInfoWindow",
-  "app/Visualization/Animation/ObjectToTable",
   "app/Visualization/Animation/Animation",
   "shims/cartodb/main"
 ], function(
@@ -11,7 +10,6 @@ define([
   Class,
   LoadingInfo,
   CartoDBInfoWindow,
-  ObjectToTable,
   Animation,
   cartodb
 ) {
